feat: calculate the amount of tokens spent on the conversation and the costs incurred

+++ internal/api-server/domain/chat/entity/biz_entity/entity_llm_chunk.go
@@ -7,15 +7,15 @@ import "github.com/lunarianss/Luna/internal/api-server/domain/chat/entity/po_ent
 // license that can be found in the LICENSE file.
 
 type LLMUsage struct {
-	PromptTokens        int     `json:"prompt_tokens"`
+	PromptTokens        int64   `json:"prompt_tokens"`
 	PromptUnitPrice     float64 `json:"prompt_unit_price"`
 	PromptPriceUnit     float64 `json:"prompt_price_unit"`
 	PromptPrice         float64 `json:"prompt_price"`
-	CompletionTokens    int     `json:"completion_tokens"`
+	CompletionTokens    int64   `json:"completion_tokens"`
 	CompletionUnitPrice float64 `json:"completion_unit_price"`
 	CompletionPriceUnit float64 `json:"completion_price_unit"`
 	CompletionPrice     float64 `json:"completion_price"`
-	TotalTokens         int     `json:"total_tokens"`
+	TotalTokens         int64   `json:"total_tokens"`
 	TotalPrice          float64 `json:"total_price"`
 	Currency            string  `json:"currency"`
 	Latency             float64 `json:"latency"`
